requirements.txt
pymongo
requests
openpyxl
python-dotenv
scikit-learn
pyyaml
xlsxwriter
streamlit

